[#17] protocol 별도 파일로 분리
